requirements.txt
google-genai
sentence-transformers
faiss-cpu
pdfplumber
streamlit
tiktoken
python-dotenv
PyMuPDF